Move main from EchoServer to ServerConsole
