+++ code/simplechat1/EchoServer.java
@@ -97,39 +97,6 @@ public class EchoServer extends AbstractServer
   synchronized protected void clientException(ConnectionToClient client, Throwable exception) {
     System.out.println("A client has unexpectedly disconnected: " + exception);
   }
-
-  //Class methods ***************************************************
-
-  /**
-   * This method is responsible for the creation of
-   * the server instance (there is no UI in this phase).
-   *
-   * @param args[0] The port number to listen on.  Defaults to 5555
-   *          if no argument is entered.
-   */
-  public static void main(String[] args)
-  {
-    int port = 0; //Port to listen on
-
-    try
-    {
-      port = Integer.parseInt(args[0]); //Get port from command line
-    }
-    catch(Throwable t)
-    {
-      port = DEFAULT_PORT; //Set port to 5555
-    }
-
-    EchoServer sv = new EchoServer(port);
-
-    try
-    {
-      sv.listen(); //Start listening for connections
-    }
-    catch (Exception ex)
-    {
-      System.out.println("ERROR - Could not listen for clients!");
-    }
-  }
 }
+
 //End of EchoServer class
